split the staticmethod into function
